sync rockchip offical sdk [buildroot]:
from (c81ae7a66f84){2019-12-25 09:21:09} package/rockchip: add linux-rga DEPENDENCIES for ffmpeg
to   (fa8f04233d32){2021-04-30 09:37:47} package: rockchip: rkadk: cp default param.ini to usr/share)

+++ buildroot/build/envsetup.sh
@@ -2,7 +2,7 @@
 
 if [ -z "${BASH_SOURCE}" ];then
 	echo Not in bash, switching to it...
-	bash -c "$0" "$@"
+	bash -c "$0 $@"
 fi
 
 function choose_board()
@@ -12,16 +12,24 @@ function choose_board()
 	echo "Lunch menu...pick a combo:"
 	echo ""
 
-	echo ${DEFCONFIG_ARRAY[@]} | xargs -n 1 | sed "=" | sed "N;s/\n/. /"
+	echo "0. non-rockchip boards"
+	echo ${RK_DEFCONFIG_ARRAY[@]} | xargs -n 1 | sed "=" | sed "N;s/\n/. /"
 
 	local INDEX
 	while true; do
-		read -p "Which would you like? [1]: " INDEX
-		INDEX=$((${INDEX:-1} - 1))
+		read -p "Which would you like? [0]: " INDEX
+		INDEX=$((${INDEX:-0} - 1))
+
+		if [ "$INDEX" -eq -1 ]; then
+			echo "Lunching for non-rockchip boards..."
+			unset TARGET_OUTPUT_DIR
+			unset RK_BUILD_CONFIG
+			break;
+		fi
 
 		if echo $INDEX | grep -vq [^0-9]; then
-			TARGET_BUILD_CONFIG="${DEFCONFIG_ARRAY[$INDEX]}"
-			[ -n "$TARGET_BUILD_CONFIG" ] && break
+			RK_BUILD_CONFIG="${RK_DEFCONFIG_ARRAY[$INDEX]}"
+			[ -n "$RK_BUILD_CONFIG" ] && break
 		fi
 
 		echo
@@ -30,33 +38,34 @@ function choose_board()
 	done
 }
 
-function lunch()
+function lunch_rockchip()
 {
-	TARGET_DIR_NAME="$TARGET_BUILD_CONFIG"
+	TARGET_DIR_NAME="$RK_BUILD_CONFIG"
 	export TARGET_OUTPUT_DIR="$BUILDROOT_OUTPUT_DIR/$TARGET_DIR_NAME"
 
 	mkdir -p $TARGET_OUTPUT_DIR || return
 
 	echo "==========================================="
 	echo
-	echo "#TARGET_BOARD=`echo $TARGET_BUILD_CONFIG | cut -d '_' -f 2`"
+	echo "#TARGET_BOARD=`echo $RK_BUILD_CONFIG | cut -d '_' -f 2`"
 	echo "#OUTPUT_DIR=output/$TARGET_DIR_NAME"
-	echo "#CONFIG=${TARGET_BUILD_CONFIG}_defconfig"
+	echo "#CONFIG=${RK_BUILD_CONFIG}_defconfig"
 	echo
 	echo "==========================================="
 
 	make -C ${BUILDROOT_DIR} O="$TARGET_OUTPUT_DIR" \
-		"$TARGET_BUILD_CONFIG"_defconfig
+		"$RK_BUILD_CONFIG"_defconfig
 
 	CONFIG=${TARGET_OUTPUT_DIR}/.config
 	cp ${CONFIG}{,.new}
+	[ -f ${CONFIG}.old ] || return 0
 	mv ${CONFIG}{.old,} &>/dev/null
 
 	make -C ${BUILDROOT_DIR} O="$TARGET_OUTPUT_DIR" olddefconfig &>/dev/null
 
 	if ! diff ${CONFIG}{,.new}; then
-		read -p "Found old config, override it? (y/n):" YES
-		[ "$YES" = "y" ] && cp ${CONFIG}{.new,}
+		read -t 10 -p "Found old config, override it? (y/n):" YES
+		[ "$YES" != "n" ] && cp ${CONFIG}{.new,}
 	fi
 }
 
@@ -67,31 +76,41 @@ function main()
 	BUILDROOT_DIR=$(dirname ${SCRIPT_DIR})
 	BUILDROOT_OUTPUT_DIR=${BUILDROOT_DIR}/output
 	TOP_DIR=$(dirname ${BUILDROOT_DIR})
-	source ${TOP_DIR}/device/rockchip/.BoardConfig.mk
 	echo Top of tree: ${TOP_DIR}
 
 	# Set croot alias
 	alias croot="cd ${TOP_DIR}"
 
-	DEFCONFIG_ARRAY=(
+	RK_DEFCONFIG_ARRAY=(
 		$(cd ${BUILDROOT_DIR}/configs/; ls rockchip_* | \
 			sed "s/_defconfig$//" | grep "$1" | sort)
 	)
 
-	DEFCONFIG_ARRAY_LEN=${#DEFCONFIG_ARRAY[@]}
-	if [ $DEFCONFIG_ARRAY_LEN -eq 0 ]; then
-		echo No available configs${1:+" for: $1"}
-		return
-	fi
+	unset RK_BUILD_CONFIG
+	RK_DEFCONFIG_ARRAY_LEN=${#RK_DEFCONFIG_ARRAY[@]}
 
-	if [ -n "$1" ]; then
-		TARGET_BUILD_CONFIG=${DEFCONFIG_ARRAY[0]}
-	else
-		choose_board
-	fi
-	[ -n "$TARGET_BUILD_CONFIG" ] || return
+	case $RK_DEFCONFIG_ARRAY_LEN in
+		0)
+			echo No available configs${1:+" for: $1"}
+			;;
+		1)
+			RK_BUILD_CONFIG=${RK_DEFCONFIG_ARRAY[0]}
+			;;
+		*)
+			if [ "$1" = ${RK_DEFCONFIG_ARRAY[0]} ]; then
+				# Prefer exact-match
+				RK_BUILD_CONFIG=$1
+			else
+				choose_board
+			fi
+			;;
+	esac
 
-	lunch
+	[ -n "$RK_BUILD_CONFIG" ] || return
+
+	source ${TOP_DIR}/device/rockchip/.BoardConfig.mk
+
+	lunch_rockchip
 }
 
 if [ "${BASH_SOURCE}" == "$0" ];then
